Finish re-organize operator module
