+++ operator/api/v1alpha2/bkapp_webhook_test.go
@@ -32,7 +32,7 @@ import (
 	paasv1alpha1 "bk.tencent.com/paas-app-operator/api/v1alpha1"
 	paasv1alpha2 "bk.tencent.com/paas-app-operator/api/v1alpha2"
 	"bk.tencent.com/paas-app-operator/pkg/config"
-	"bk.tencent.com/paas-app-operator/pkg/utils/kubetypes"
+	"bk.tencent.com/paas-app-operator/pkg/kubeutil"
 	"bk.tencent.com/paas-app-operator/pkg/utils/stringx"
 )
 
@@ -647,7 +647,7 @@ var _ = Describe("test webhook.Validator", func() {
 		It("Normal", func() {
 			legacyProcResConfig := make(paasv1alpha2.LegacyProcConfig)
 			legacyProcResConfig["web"] = map[string]string{"cpu": "2", "memory": "2G"}
-			_ = kubetypes.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
+			_ = kubeutil.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
 
 			err := bkapp.ValidateCreate()
 			Expect(err).To(BeNil())
@@ -655,7 +655,7 @@ var _ = Describe("test webhook.Validator", func() {
 		It("Invalid unset", func() {
 			legacyProcResConfig := make(paasv1alpha2.LegacyProcConfig)
 			legacyProcResConfig["web"] = map[string]string{"cpu": "", "memory": "2G"}
-			_ = kubetypes.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
+			_ = kubeutil.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
 
 			err := bkapp.ValidateCreate()
 			Expect(err).NotTo(BeNil())
@@ -663,7 +663,7 @@ var _ = Describe("test webhook.Validator", func() {
 		It("Invalid exceed cpu max limit", func() {
 			legacyProcResConfig := make(paasv1alpha2.LegacyProcConfig)
 			legacyProcResConfig["web"] = map[string]string{"cpu": "6", "memory": "2G"}
-			_ = kubetypes.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
+			_ = kubeutil.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
 
 			err := bkapp.ValidateCreate()
 			Expect(err).NotTo(BeNil())
@@ -671,7 +671,7 @@ var _ = Describe("test webhook.Validator", func() {
 		It("Invalid exceed memory max limit", func() {
 			legacyProcResConfig := make(paasv1alpha2.LegacyProcConfig)
 			legacyProcResConfig["web"] = map[string]string{"cpu": "2", "memory": "8G"}
-			_ = kubetypes.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
+			_ = kubeutil.SetJsonAnnotation(bkapp, paasv1alpha2.LegacyProcResAnnoKey, legacyProcResConfig)
 
 			err := bkapp.ValidateCreate()
 			Expect(err).NotTo(BeNil())
