--- operator/pkg/utils/kubestatus/health.go
@@ -1,33 +1,0 @@
-/*
- * TencentBlueKing is pleased to support the open source community by making
- * 蓝鲸智云 - PaaS 平台 (BlueKing - PaaS System) available.
- * Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
- * Licensed under the MIT License (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- *	http://opensource.org/licenses/MIT
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * We undertake not to change the open source license (MIT license) applicable
- * to the current version of the project delivered to anyone in the future.
- */
-
-package kubestatus
-
-import paasv1alpha2 "bk.tencent.com/paas-app-operator/api/v1alpha2"
-
-// HealthStatus the resource health status
-type HealthStatus struct {
-	Phase   paasv1alpha2.HealthPhase
-	Reason  string
-	Message string
-}
-
-func (s *HealthStatus) withMessage(message string) *HealthStatus {
-	s.Message = message
-	return s
-}
